Extract cutlayout logic from worker into it's own file. First template of how I'd like to extract most logic into dedicated files instead of the worker.js megafile
