+++ src/App.jsx
@@ -14,7 +14,7 @@ import GlobalVariables from "./js/globalvariables.js";
 import LoginMode from "./components/main-routes/LoginMode.jsx";
 import RunMode from "./components/main-routes/RunMode.jsx";
 import CreateMode from "./components/main-routes/CreateMode.jsx";
-import cadWorker from "./worker.js?worker";
+import cadWorker from "./worker/worker.js?worker";
 import { button } from "leva";
 import { QueryClient, QueryClientProvider } from "react-query";
 import Callback from "./components/main-routes/CallBack.jsx";
